mysite/hotel_app/permissions.py
from  rest_framework import  permissions

class CheckStatusUser(permissions.BasePermission):
    def has_permission(self, request, view):
        if request.method in permissions.SAFE_METHODS:# get,
            return  True
        elif request.user.status == 'owner':
            return True


class CheckOwner(permissions.BasePermission):
     def has_object_permission(self, request, view, obj):
         if request.method in permissions.SAFE_METHODS:
             return True
         elif request.user == obj.owner:
             return True